GH-6 Moving intents, creating test folders
